update coord conversion to match WMM technical report
While trying to complete future steps I found it difficult to carry on
with the coordinate system I had.

I had the idea to conver to spherical when I realized the WMM used
spherical harmonics but was unsure if I was going about this correctly
and if I should even use the spherical harmonics.

When I read the WMM Technical Report it seems this is the correct idea.
Updated coordinate transfer function to match.