update api to return lat lon
sort of stuck here because this seems like a really silly way to proceed

on the FE we're using cartesian space, converting to spherical
(lat/lon), sending it to the BE, calculating, sending back, and then
converting to cartesian???

there's gotta be a better way. if we can start from spherical that could
make a world (lol) of difference.

obviously, we wouldn't even want to start from spherical but rather
geodetic so that's another wrench to think about....